+++ tp-progra2/src/trabajo/Paquete.java
@@ -2,12 +2,12 @@ package trabajo;
 
 public class Paquete {
 	
-	private int peso;
-	private int volumen;
+	private double peso;
+	private double volumen;
 	private String destino;
 	private boolean refrigeracion;
 	
-	public Paquete (int peso, int volumen, String destino, boolean refrigeracion) {
+	public Paquete (double peso, double volumen, String destino, boolean refrigeracion) {
 		
 		this.peso = peso;
 		this.volumen = volumen;
